clubbed-svc-enquiry-handler-2: Details sent in body rather that URL. Uses Async handling. Improved validation and error reporting
First and last name are sent in the body of the message instead of the URL.

Given that commands execute asynchronously, the call uses the async dispatcher feature to return a deferred result. When the command executes, a success or failure response can be returned to the client.

The user details undergo stricter validation and bind errors are summarised and returned to the client for easier error reporting