clubbed-svc-enquiry-handler-3: Designed to be run as multiple instances, one per domain
Runs as a standalone Spring Boot application. Persistence is to a specific schema per domain. Published event to a specific queue by providing a routing key that is the queues name. This allows multiple instances of the application to run, one per domain. Consumes it's configuration from Spring Config Server

Fixes a bug that could allow a new enquiry event to be published even if creating the Enquiry aggregate failed due to a unique constraint violation.